pyproject.toml
[project]
name = "main-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pillow>=12.2.0",
    "torch<2.6",
    "torchvision>=0.20.1",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu118"

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu121"
